refactor: 重命名实体抽象类
