+++ src/index.ts
@@ -1,4 +1,4 @@
 export { Mapping } from './annotations/property/mapping'
 export { NotBeNull, NotBeEmpty, NotBeUndefined } from './annotations/property/asserts'
 
-export { Entity } from './models/Entity'
+export { DataEntity } from './models/DataEntity'
