+++ src/models/isDataEntity.ts
