+++ test/serialize.spec.ts
@@ -1,6 +1,6 @@
 /// <reference types="../typings/model" />
 
-import { Entity, Mapping, NotBeEmpty, NotBeNull, NotBeUndefined } from '../src'
+import { DataEntity, Mapping, NotBeEmpty, NotBeNull, NotBeUndefined } from '../src'
 import { MetadataContext } from '../src/metadata/MetadataContext'
 
 describe('序列化', () => {
@@ -10,7 +10,7 @@ describe('序列化', () => {
 
   describe('@Mapping()', () => {
     it('序列化实例成同构数据', () => {
-      class ResponseData<T> extends Entity {
+      class ResponseData<T> extends DataEntity {
         @Mapping()
         public data?: T
 
@@ -36,7 +36,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例成异构数据', () => {
-      class ResponseData<T> extends Entity {
+      class ResponseData<T> extends DataEntity {
         @Mapping()
         public data?: T
 
@@ -54,7 +54,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到同构复合结构数据', () => {
-      class Pattern extends Entity {
+      class Pattern extends DataEntity {
         @Mapping()
         public id?: number
 
@@ -62,7 +62,7 @@ describe('序列化', () => {
         public strategy?: Strategy
       }
 
-      class Strategy extends Entity {
+      class Strategy extends DataEntity {
         @Mapping()
         name?: string
 
@@ -70,7 +70,7 @@ describe('序列化', () => {
         metric?: Metric
       }
 
-      class Metric extends Entity {
+      class Metric extends DataEntity {
         @Mapping({ path: 'jar' })
         bar?: string
       }
@@ -90,7 +90,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到同构复合结构数组', () => {
-      class LogSource extends Entity {
+      class LogSource extends DataEntity {
         @Mapping()
         name?: string
 
@@ -98,7 +98,7 @@ describe('序列化', () => {
         categories?: Category[]
       }
 
-      class Category extends Entity {
+      class Category extends DataEntity {
         @Mapping()
         name?: string
       }
@@ -121,17 +121,17 @@ describe('序列化', () => {
     })
 
     it('序列化实例到复合结构数组的特定位置 [n]', () => {
-      class LogSource extends Entity {
+      class LogSource extends DataEntity {
         @Mapping()
         name?: string
       }
 
-      class Category extends Entity {
+      class Category extends DataEntity {
         @Mapping()
         name?: string
       }
 
-      class Rule extends Entity {
+      class Rule extends DataEntity {
         // logSource 使用 filters 的第 0 位置
         @Mapping({ relatedEntityDescriptor: 'LogSource', path: 'filters[0]' })
         logSource?: LogSource
@@ -156,17 +156,17 @@ describe('序列化', () => {
     })
 
     it('序列化实例到复合结构数组的特定位置 [n][m]', () => {
-      class LogSource extends Entity {
+      class LogSource extends DataEntity {
         @Mapping()
         name?: string
       }
 
-      class Category extends Entity {
+      class Category extends DataEntity {
         @Mapping()
         name?: string
       }
 
-      class Rule extends Entity {
+      class Rule extends DataEntity {
         @Mapping({ relatedEntityDescriptor: 'LogSource', path: 'filters[0][1]' })
         logSource?: LogSource
 
@@ -189,7 +189,7 @@ describe('序列化', () => {
     })
 
     it('序列化递归数据结构', () => {
-      class Pattern extends Entity {
+      class Pattern extends DataEntity {
         @Mapping()
         id?: number
 
@@ -216,7 +216,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到混合切片数据 [n:m]', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping({ path: 'filters[0]' })
         id?: number
 
@@ -242,7 +242,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到切片数据 [:]', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping({ path: 'filters[:]' })
         filters?: number[]
       }
@@ -261,7 +261,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到切片数据 []', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping({ path: 'filters[]' })
         filters?: number[]
       }
@@ -280,7 +280,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到切片数据 [n:]', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping({ path: 'filters[2:]' })
         filters?: number[]
       }
@@ -301,7 +301,7 @@ describe('序列化', () => {
     })
 
     it('序列化实例到切片数据 [:m]', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping({ path: 'filters[:4]' })
         filters?: number[]
       }
@@ -320,7 +320,7 @@ describe('序列化', () => {
     })
 
     it('序列化间接继承 Entity 的类型实例', () => {
-      class Foo extends Entity {
+      class Foo extends DataEntity {
         @Mapping()
         id?: number
 
@@ -354,7 +354,7 @@ describe('序列化', () => {
     })
 
     it('序列化支持symbol数据类型', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @Mapping({path: 'idList'})
         [Symbol.iterator]?: string[]
       }
@@ -367,7 +367,7 @@ describe('序列化', () => {
     })
 
     it('自定义序列化过程', () => {
-      class Company extends Entity {
+      class Company extends DataEntity {
         constructor() {
           super()
 
@@ -390,7 +390,7 @@ describe('序列化', () => {
 
   describe('@NotBeNull()', () => {
     it('默认错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeNull()
         @Mapping()
         id?: number | null = null
@@ -404,7 +404,7 @@ describe('序列化', () => {
     })
 
     it('自定义错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeNull('忘了写id, 忘了写id, 忘了写id, 重要的事情说三遍!')
         @Mapping()
         id?: number | null = null
@@ -420,7 +420,7 @@ describe('序列化', () => {
 
   describe('@NotBeUndefined()', () => {
     it('默认错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeUndefined()
         @Mapping()
         id?: number
@@ -434,7 +434,7 @@ describe('序列化', () => {
     })
 
     it('自定义错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeUndefined('忘了写id, 忘了写id, 忘了写id, 重要的事情说三遍!')
         @Mapping()
         id?: number
@@ -450,7 +450,7 @@ describe('序列化', () => {
 
   describe('@NotBeEmpty()', () => {
     it('默认错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeEmpty()
         @Mapping()
         id?: number | null = null
@@ -464,7 +464,7 @@ describe('序列化', () => {
     })
 
     it('自定义错误信息', () => {
-      class Employee extends Entity {
+      class Employee extends DataEntity {
         @NotBeEmpty('id 不能为 empty')
         @Mapping()
         id?: number | null = null
